Add ps1b.py
Have the user input a semi-annual salary raise semi_annual_raise (as a decimal percentage)
2. After the 6th month, increase your salary by that percentage. Do the same after the 12th month, the 18th month, and so on.
Write a program to calculate how many months it will take you save up enough money for a down
payment. Like before, assume that your investments earn a return of r = 0.04 (or 4%) and the
required down payment percentage is 0.25 (or 25%). Have the user enter the following variables:
1. The starting annual salary (annual_salary)2
2. The percentage of salary to be saved (portion_saved)
3. The cost of your dream home (total_cost)
4. The semiannual salary raise (semi_annual_raise)